Prefabs and Pivot Points
A system for entity prefab scripts now exists. This allows for quick
templating of entities, especially multiple entities of similar
structure. These prefab instances can be expanded after being built too.
There is also the addition of pivot points, which all for more control
of the positioning of rectangles and sprites. Text pivoting to come
soon, even though it is harder to do.