Framerate Counter
A framerate counter nature has been added to the engine. To get this to
work, some additional restructuring was needed. sc_main now has a header
file, letting its globals and macros be shared with other files. A
function for float to string has also been added, with a simplified
function name convention for these types of methods (still a bad
solution all around, but someday). This concludes the UI rendering
branch.