On the Path to Getting Layered, Orthographic UI Rendering
Right now this feature is hampered by a nasty vtable compilation issue.
However, on the path to getting it resolved a large amount of
refactoring has happened, such as the addition of tagging components, a
far better method for maintaining pointer lists to certain component
categories, and big decoupling changes for State, which is leading to
more clear design.