Rewrite to use New/Delete and Orthographic Layers
Assets and components are now handled in memory completely differently.
Instead of allocating large swathes of memory for particular assets and
components, their memory is reserved individually via new. This allowed
certain memory leak issues to be resolved, including the vtables leak
issue which was preventing the sorting of orthographic draw components.
This has now all been fixed and 2D screen elements can be sorted by the
layer they are assigned to. Pooling may return in a different form if
performance issues require it.