Beginning of Buttons and New Stage Data Format, Plus Better Tokenizer
On the path to add stage loading and saving, basic button rendering has
been added. Stages are now stored in a 1D format for easier compression.
This more fleshed out stage format required the creation of a much
better (and correct) tokenizer. A fresh new stage can be created without
having loaded in a stage to begin with. Newer, better systems for
loading in default textures and brushes will be needed in the future.