Transform Improvements
Transforms are now in a hierarchy. This is maintained efficiently with a
dirty flag system; only transforms that need to be recalculated are.
This makes working with transforms way simpler; instead of having to
recalculate the world matrix manually, the function for such is
automatically called in the background per frame. Robust getter and
setter functions have been added to Transform to make all of this
invisible. This should make building GUI's much easier, but will also to
contribute when more complicated models and game objects get introduced.