Component Activation
Stolen shamelessly from Unity. You can toggle any component on and off.
Certain components behave differently when turned off (like draws not
rendering and natures not updating). If a transform is turned on/off,
all the components in the object are turned on/off. A example editor
shortcut is included in this commit only. Will be used first to turn GUI
components on/off.