Further Corrections to Transform Tree
Orthographic Draw components now fully use their transform instead of
storing their own seperate size values. Fixes to the transform
parent-child hierarchy so that it works better with UI elements as well
as world elements. Children can choose to not inherit a part of their
parent's transform (in the case of UI, not usually inheriting scale).
Not the best solution, but works for the time being. Might be too much
system for a single case. Many minor tweaks as well. Editor UI Panel
prefab was added. UI Prefabs now have layer as an argument.