Added EditorTextField, Text Input, and Naming Stages
Input now supports text input from the keyboard into a string buffer
that can be placed where Nature's want it. Not tested for security,
though there is a little bit of an assumption that SDL handles at least
a little bit of that (maybe not). This is used by the EditorTextField
prefab, which is a traditional text input GUI element. As a
demonstration of this, the ability to title a stage has been added to
the editor.