Transform Refactoring
The Transform component class is now more nicely written. A hopefully
better way of doing global transform components has been added.  Only
transforms that are given changes in a frame are queued for calculation.
Transforms are automatically added to every entity and cannot be
removed. This way all entities can be part of hierarchies. Code cleaning
done to both State and Transform algorithms.